vertex-ai/pipelines/pyproject.toml
[tool.poetry]
name = "vertex-ai-pipelines-scratchpad"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "vertex_ai_pipelines_scratchpad"}]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-aiplatform = "^1.24.0"
transparentpath = "^1.1.38"
kfp = "1.8.9"
google-cloud-storage = "<2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
